sprucing up visits

+++ app/controllers/UnhlsPatientController.php
@@ -190,15 +190,21 @@ class UnhlsPatientController extends \BaseController {
 	 */
 	public function delete($id)
 	{
-		//Soft delete the patient
+		// if no visit made, soft delete
 		$patient = UnhlsPatient::find($id);
 
-		$patient->delete();
-
+		$patientInUse = UnhlsVisit::where('patient_id', '=', $id)->first();
+		if (empty($patientInUse)) {
+			// The has no visit
+			$patient->delete();
+		} else {
+			// The has visit
+			return Redirect::route('unhls_patient.index')
+				->with('message', 'This Patient has visits, not Deleted!');
+		}
 		// redirect
-			$url = Session::get('SOURCE_URL');
-			return Redirect::to($url)
-			->with('message', 'The commodity was successfully deleted!');
+		return Redirect::route('unhls_patient.index')
+			->with('message', 'Patient Successfully Deleted!');
 	}
 
 	/**
@@ -210,29 +216,4 @@ class UnhlsPatientController extends \BaseController {
 	{
         return UnhlsPatient::search(Input::get('text'))->take(Config::get('kblis.limit-items'))->get()->toJson();
 	}
-
-	/**
-	 *Return a unique Lab Number
-	 *
-	 * @return string of current age concatenated with incremental Number.
-	 */
-	// Private function generateUniqueLabID(){
-
-	// 	//Get Year, Month and day of today. If Jan O1 then reset last insert ID to 1 to start a new cycle of IDs
-	// 	$year = date('Y');
-	// 	$month = date('m');
-	// 	$day = date('d');
-
-	// 	if($month == '01' && $day == '01'){
-	// 		$lastInsertId = 1;
-	// 	}
-	// 	else{
-	// 		$lastInsertId = DB::table('unhls_patients')->max('id')+1;
-	// 	}
-	// 	$fcode = \Config::get('constants.FACILITY_CODE');
-	// 	$num = $year.str_pad($lastInsertId, 6, '0', STR_PAD_LEFT);
-	// 	return $fcode.'-'.$num;
-	// }
-
-
 }
